+++ Assets/Scripts/Class/Plant4.cs
@@ -4,7 +4,7 @@ using UnityEngine;
 
 public class Plant4 : PlantBase
 {
-    public Plant4(string plantName, string plantType, int growthRate, float hydration, float nutrition, bool isSick, bool isThirsty)
+    public Plant4(string plantName, string plantType, int growthRate, int hydration, int nutrition, bool isSick, bool isThirsty)
     {
         this.plantName = plantName;
         this.plantType = plantType;
